pyproject.toml
[project]
name = "lore-forger"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.0",
    "google-cloud-firestore>=2.23.0",
    "google-genai>=1.65.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
